feat: move middleware to train/ and support train/train number urls
